Update App to use new MinisterIssues component
